requirements.txt
aiogram==3.13.0
google-generativeai==0.8.3
matplotlib>=3.7
python-dotenv==1.0.1
aiohttp==3.10.10
asyncpg>=0.29.0
